Changes for Rosa, Best Charge State with Priming, Matched Trigger/Target Lists
After a discussion with Rosa I removed the "Both" column from the
modifications, while also overhauling how I deal with modifications.
Both dynamic and static modifications are stored in one large, ugly,
dictionary. It is the most effective way to shuttle them around. I also
implemented a check to target the best charge state if a raw file is
provided with the target list and no charge state is provided. Also, I
now ensure that the trigger and target lists are identical, which means
that we will not be using non-TMT containing ions for identification.
Lastly, I now carry the charge information all the way to the raw file
create for any target mass list.