Fixed bug with analysis
There was a bug with creating the modification dictionary that crashed
the program when analyzing. There is a known issue when the number of
ions for triggering is less than the value in the method. I will need to
fix this somehow...probably add in dummy ions with masses that are very
close so that the number is equal to the cutoff.